remove getExhibitionAndEventPromos
