+++ server/index.js
@@ -13,7 +13,6 @@ import {
   getGlobalAlert,
   getPaginatedEventPromos,
   getPaginatedExhibitionPromos,
-  getExhibitionAndEventPromos,
   convertPrismicResultsToPaginatedResults,
   createEventPromos,
   createIndividualEventPromos,
@@ -41,7 +40,6 @@ export const prismic = {
   asHtml,
   getPaginatedEventPromos,
   getPaginatedExhibitionPromos,
-  getExhibitionAndEventPromos,
   convertPrismicResultsToPaginatedResults,
   createEventPromos,
   createIndividualEventPromos,
