+++ server/services/prismic.js
@@ -48,9 +48,6 @@ const eventFields = [
   'organisations.name', 'organisations.image',
   'organisations.url', 'background-textures.image'
 ];
-const exhibitionFields = [
-  'exhibition-formats.title'
-];
 
 export const defaultPageSize = 40;
 
@@ -370,219 +367,4 @@ export async function getPaginatedExhibitionPromos(page: number): Promise<Exhibi
   const promos = createExhibitionPromos(exhibitions.results);
   const paginatedResults = convertPrismicResultsToPaginatedResults(exhibitions);
   return paginatedResults(promos);
-}
-
-// Returns true if the date range of an event coincides with the date range provided
-//                    [_____date range_____]
-//          [___event1___]              [___event2___]
-//                         [___event3___]
-function datesOverlapRange (eventStartDate, eventEndDate, rangeStartDate, rangeEndDate) {
-  const eventStart = london(eventStartDate);
-  const eventEnd = london(eventEndDate);
-  const rangeStart = rangeStartDate && london(rangeStartDate);
-  const rangeEnd = rangeEndDate && london(rangeEndDate);
-
-  return (rangeEnd === undefined || eventStart.isSame(rangeEnd, 'day') || eventStart.isBefore(rangeEnd, 'day')) && (eventEnd.isSame(rangeStart, 'day') || eventEnd.isAfter(rangeStart, 'day'));
-}
-
-function filterPromosByDate(promos, startDate, endDate) {
-  return promos.filter(e => datesOverlapRange(e.start, e.end, startDate, endDate));
-}
-
-function filterCurrentExhibitions(promos, todaysDate) {
-  return promos.filter((e) => {
-    const eventStart = e.start && london(e.start);
-    const eventEnd = e.end && london(e.end);
-
-    return eventEnd === null || todaysDate.isSame(eventStart, 'day') || todaysDate.isSame(eventEnd, 'day') || todaysDate.isBefore(eventEnd, 'day') && todaysDate.isAfter(eventStart, 'day');
-  });
-}
-
-function filterUpcomingExhibitions(promos, todaysDate) {
-  return promos.filter((e) => {
-    const eventStart = london(e.start);
-    return todaysDate.isBefore(eventStart, 'day');
-  });
-}
-
-function getActiveState(today, fromDate, toDate) {
-  const rangeStart = fromDate && london(fromDate);
-  const rangeEnd = toDate && london(toDate);
-  if (rangeStart && !rangeEnd) {
-    return 'everything';
-  } else if (today.isSame(rangeStart, 'day') && today.isSame(rangeEnd, 'day')) {
-    return 'today';
-  } else if (rangeStart.isSame(getWeekendFromDate(today), 'day') && rangeEnd.isSame(getWeekendToDate(today), 'day')) {
-    return 'weekend';
-  }
-};
-
-// On the 'everything' view, events are displayed per month
-// If an event has dates that span several months (and/or years) it needs to appear for each month
-// This function duplicates an event for each month in which it appears and groups the events into the relevant months
-function duplicatePromosByMonthYear(promos) {
-  return promos.reduce((acc, currEvent) => {
-    const start = london(currEvent.start);
-    const end = london(currEvent.end);
-    if (end.isSame(start, 'month')) {
-      const year = start.format('YYYY');
-      const month = start.format('MMMM');
-      if (!acc[year]) {
-        acc[year] = {};
-      }
-      if (!acc[year][month]) {
-        acc[year][month] = [];
-      }
-      acc[year][month].push(currEvent);
-    } else {
-      while (end.isAfter(start) || end.isSame(start, 'month')) {
-        const year = start.format('YYYY');
-        const month = start.format('MMMM');
-        if (!acc[year]) {
-          acc[year] = {};
-        }
-        if (!acc[year][month]) {
-          acc[year][month] = [];
-        }
-        acc[year][month].push(currEvent);
-        start.add(1, 'month');
-      }
-    }
-    return acc;
-  }, {});
-}
-
-function getListHeader(dates, collectionOpeningTimes) {
-  const todaysDate = london().startOf('day');
-  const todayString = todaysDate.format('dddd');
-  const galleryOpeningHours = collectionOpeningTimes.placesOpeningHours && collectionOpeningTimes.placesOpeningHours.find(venue => venue.name === 'Galleries').openingHours;
-  const regularOpeningHours = galleryOpeningHours && galleryOpeningHours.regular.find(i => i.dayOfWeek === todayString);
-  const exceptionalOpeningHours = galleryOpeningHours && galleryOpeningHours.exceptional && galleryOpeningHours.exceptional.find(i => {
-    const dayOfWeek = london(i.overrideDate).startOf('day');
-    return todaysDate.isSame(dayOfWeek);
-  });
-  const todayOpeningHours = exceptionalOpeningHours || regularOpeningHours;
-
-  return {
-    todayOpeningHours,
-    name: 'What\'s on',
-    items: [
-      {
-        id: 'everything',
-        title: 'Everything',
-        url: `/whats-on`
-      },
-      {
-        id: 'today',
-        title: 'Today',
-        url: `/whats-on/today`
-      },
-      {
-        id: 'the-weekend',
-        title: 'This weekend',
-        url: `/whats-on/the-weekend`
-      }
-    ]
-  };
-}
-
-function getMomentsForPeriod(period) {
-  const todaysDate = london();
-  const todaysDatePlusSix = todaysDate.clone().add(6, 'days');
-
-  switch (period) {
-    case 'today': return [todaysDate.startOf('day'), todaysDate.endOf('day')];
-    case 'this-weekend': return [getWeekendFromDate(todaysDate), getWeekendToDate(todaysDate)];
-    // FIXME: this isn't really 'this week', but the 'next seven days' (needs UX/content rethink?)
-    case 'this-week': return [todaysDate.startOf('day'), todaysDatePlusSix.endOf('day')];
-    default: return [todaysDate.startOf('day'), undefined];
-  }
-}
-
-export async function getExhibitionAndEventPromos(dateRange, collectionOpeningTimes) {
-  const todaysDate = london();
-  const [fromDateMoment, toDateMoment] = getMomentsForPeriod(dateRange);
-  const fromDate = fromDateMoment.format('YYYY-MM-DD');
-  const toDate  = toDateMoment ? toDateMoment.format('YYYY-MM-DD') : undefined;
-
-  // Had to split exhibitions and installations query from events query.
-  // We can't query events by dateAfter and there is a max of 100 results per query.
-  // Exhibitions would start disappearing, once results started goind over 100 - we were getting close.
-  const allExhibitionsAndInstallationsPromise = getAllOfType(['exhibitions', 'installations'], {
-    pageSize: 100,
-    fetchLinks: exhibitionFields,
-    orderings: '[my.exhibitions.start]'
-  });
-  const allEventsPromise = getAllOfType(['events'], {
-    pageSize: 100,
-    fetchLinks: eventFields,
-    orderings: '[my.events.times.startDateTime desc]'
-  });
-
-  const [ allExhibitionsAndInstallations, allEvents ] = await Promise.all([allExhibitionsAndInstallationsPromise, allEventsPromise]);
-
-  const exhibitionPromos = createExhibitionPromos(allExhibitionsAndInstallations.results.filter(e => e.type === 'exhibitions'));
-  const permanentExhibitionPromos = filterCurrentExhibitions(exhibitionPromos.filter(e => e.format && e.format.title.toLowerCase() === 'permanent'), todaysDate);
-  const temporaryExhibitionPromos = filterPromosByDate(exhibitionPromos.filter(e => !e.format || e.format && e.format.title.toLowerCase() !== 'permanent'), fromDate, toDate);
-  const currentTemporaryExhibitionPromos = filterCurrentExhibitions(temporaryExhibitionPromos, todaysDate);
-  const upcomingTemporaryExhibitionPromos = filterUpcomingExhibitions(temporaryExhibitionPromos, todaysDate);
-  const eventPromos = filterPromosByDate(createEventPromos(allEvents.results), fromDate, toDate).sort((a, b) => a.start.localeCompare(b.start));
-  const individualEventPromos = filterPromosByDate(createIndividualEventPromos(allEvents.results), fromDate, toDate).sort((a, b) => a.start.localeCompare(b.start));
-
-  // TODO: get rid of this snowflake when what's on/homepage are capable of handling manual lists.
-  const colourOfPharmacyPromo = createExhibitionPromos(allExhibitionsAndInstallations.results.filter(e => e.id === 'WyuWLioAACkACeYv'))[0];
-
-  // eventPromosSplitAcrossMonths and monthControls only required for the 'everything' view
-  const eventPromosSplitAcrossMonths = duplicatePromosByMonthYear(eventPromos);
-  const monthControls = Object.keys(eventPromosSplitAcrossMonths).reduce((acc, year) => {
-    Object.keys(eventPromosSplitAcrossMonths[year]).forEach((month) => {
-      const controlObject = {
-        id: `${year}-${month}`,
-        text: month,
-        url: `#${month}-${year}`
-      };
-      acc.push(controlObject);
-    });
-    return acc;
-  }, []);
-  const dates = {
-    today: todaysDate.format('YYYY-MM-DD'),
-    weekend: [getWeekendFromDate(todaysDate).format('YYYY-MM-DD'), getWeekendToDate(todaysDate).format('YYYY-MM-DD')],
-    all: [todaysDate.format('YYYY-MM-DD')],
-    queriedDates: [fromDate, toDate]
-  };
-  const active = getActiveState(todaysDate, fromDate, toDate);
-  const listHeader = getListHeader(dates, collectionOpeningTimes);
-  return {
-    active,
-    dates,
-    permanentExhibitionPromos,
-    temporaryExhibitionPromos,
-    colourOfPharmacyPromo,
-    currentTemporaryExhibitionPromos,
-    upcomingTemporaryExhibitionPromos,
-    eventPromos,
-    individualEventPromos,
-    eventPromosSplitAcrossMonths,
-    monthControls,
-    listHeader
-  };
-}
-
-function getWeekendFromDate(today) {
-  const todayInteger = today.day(); // day() return Sun as 0, Sat as 6
-  if (todayInteger !== 0) {
-    return london(today).day(5);
-  } else {
-    return london(today).day(-2);
-  }
-}
-
-function getWeekendToDate(today) {
-  const todayInteger = today.day(); // day() return Sun as 0, Sat as 6
-  if (todayInteger === 0) {
-    return london(today);
-  } else {
-    return london(today).day(7);
-  }
 }
