+++ whats_on/app/views/partials/date_range.njk
@@ -1,21 +1,21 @@
-{% set fromDate = exhibitionAndEventPromos.dates.queriedDates[0] %}
-{% set toDate = exhibitionAndEventPromos.dates.queriedDates[1] %}
+{% set fromDate = dateRange[0] %}
+{% set toDate = dateRange[1] %}
 
 <p class="{{- [
   {s:'HNM4', m:'HNM4'} | fontClasses,
   {s:0} | spacingClasses({margin: ['top']}),
   {s:1} | spacingClasses({margin: ['bottom']})
 ].join(' ') -}}">
-  {% if exhibitionAndEventPromos.active == 'today' %}
+  {% if period == 'today' %}
     <time datetime="{{ fromDate }}">{{ fromDate | formatDate }}</time>
-  {% elseif exhibitionAndEventPromos.active == 'weekend' %}
+  {% elseif period == 'this-weekend' %}
     <time datetime="{{ fromDate }}">{{ fromDate | formatDay }}</time>&ndash;<time datetime="{{ toDate }}">{{ toDate | formatDay }}</time>
-  {% elseif exhibitionAndEventPromos.active == 'everything' %}
+  {% elseif period == 'current-and-coming-up' %}
     From <time datetime="{{ fromDate }}">{{ fromDate | formatDate }}</time>
   {% endif %}
 </p>
 
-{% if not exhibitionAndEventPromos.listHeader.todayOpeningHours.opens and exhibitionAndEventPromos.active === 'today' %}
+{% if not listHeader.todayOpeningHours.opens and period === 'today' %}
   <p class="{{ {s:'WB6', m:'WB5', l:'WB4'} | fontClasses }} {{{s:2} | spacingClasses({margin: ['bottom']})}}">Our exhibitions are closed today, but our <a href="{{cafePromo.url}}">café</a> and <a href="{{shopPromo.url}}">shop</a> are open for your visit.</p>
   <div class="{{ {s:3, m:4, l:5} | spacingClasses({margin: ['top', 'bottom']}) }}">
     {% componentV2 'divider', null, {'dashed': true} %}
